requirements.txt
aiogram==3.5.0
python-dotenv==1.0.1
pandas==2.2.2
numpy==1.26.4
ta==0.11.0
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.2.1
asyncio==3.4.3
websockets==12.0
selenium==4.20.0
webdriver-manager==4.0.1
pytz==2024.1
telethon==1.35.0
psutil==5.9.8
nest-asyncio==1.6.0
aiohttp==3.9.5
aiofiles==23.2.1
loguru==0.7.2
BinaryOptionsToolsV2==0.1.5